+++ packages/flutter/lib/src/native/java/binding.dart
@@ -1305,6 +1305,83 @@ class SentryFlutterPlugin$Companion extends jni$_.JObject {
         .object<ReplayIntegration?>(const $ReplayIntegration$NullableType());
   }
 
+  static final _id_getApplicationContext = _class.instanceMethodId(
+    r'getApplicationContext',
+    r'()Landroid/content/Context;',
+  );
+
+  static final _getApplicationContext = jni$_.ProtectedJniExtensions.lookup<
+          jni$_.NativeFunction<
+              jni$_.JniResult Function(
+                jni$_.Pointer<jni$_.Void>,
+                jni$_.JMethodIDPtr,
+              )>>('globalEnv_CallObjectMethod')
+      .asFunction<
+          jni$_.JniResult Function(
+            jni$_.Pointer<jni$_.Void>,
+            jni$_.JMethodIDPtr,
+          )>();
+
+  /// from: `public final android.content.Context getApplicationContext()`
+  /// The returned object must be released after use, by calling the [release] method.
+  jni$_.JObject? getApplicationContext() {
+    return _getApplicationContext(
+            reference.pointer, _id_getApplicationContext as jni$_.JMethodIDPtr)
+        .object<jni$_.JObject?>(const jni$_.JObjectNullableType());
+  }
+
+  static final _id_loadContextsAsBytes = _class.instanceMethodId(
+    r'loadContextsAsBytes',
+    r'()[B',
+  );
+
+  static final _loadContextsAsBytes = jni$_.ProtectedJniExtensions.lookup<
+          jni$_.NativeFunction<
+              jni$_.JniResult Function(
+                jni$_.Pointer<jni$_.Void>,
+                jni$_.JMethodIDPtr,
+              )>>('globalEnv_CallObjectMethod')
+      .asFunction<
+          jni$_.JniResult Function(
+            jni$_.Pointer<jni$_.Void>,
+            jni$_.JMethodIDPtr,
+          )>();
+
+  /// from: `public final byte[] loadContextsAsBytes()`
+  /// The returned object must be released after use, by calling the [release] method.
+  jni$_.JByteArray? loadContextsAsBytes() {
+    return _loadContextsAsBytes(
+            reference.pointer, _id_loadContextsAsBytes as jni$_.JMethodIDPtr)
+        .object<jni$_.JByteArray?>(const jni$_.JByteArrayNullableType());
+  }
+
+  static final _id_loadDebugImagesAsBytes = _class.instanceMethodId(
+    r'loadDebugImagesAsBytes',
+    r'(Ljava/util/Set;)[B',
+  );
+
+  static final _loadDebugImagesAsBytes = jni$_.ProtectedJniExtensions.lookup<
+              jni$_.NativeFunction<
+                  jni$_.JniResult Function(
+                      jni$_.Pointer<jni$_.Void>,
+                      jni$_.JMethodIDPtr,
+                      jni$_.VarArgs<(jni$_.Pointer<jni$_.Void>,)>)>>(
+          'globalEnv_CallObjectMethod')
+      .asFunction<
+          jni$_.JniResult Function(jni$_.Pointer<jni$_.Void>,
+              jni$_.JMethodIDPtr, jni$_.Pointer<jni$_.Void>)>();
+
+  /// from: `public final byte[] loadDebugImagesAsBytes(java.util.Set<java.lang.String> set)`
+  /// The returned object must be released after use, by calling the [release] method.
+  jni$_.JByteArray? loadDebugImagesAsBytes(
+    jni$_.JSet<jni$_.JString> set,
+  ) {
+    final _$set = set.reference;
+    return _loadDebugImagesAsBytes(reference.pointer,
+            _id_loadDebugImagesAsBytes as jni$_.JMethodIDPtr, _$set.pointer)
+        .object<jni$_.JByteArray?>(const jni$_.JByteArrayNullableType());
+  }
+
   static final _id_new$ = _class.constructorId(
     r'(Lkotlin/jvm/internal/DefaultConstructorMarker;)V',
   );
@@ -1686,6 +1763,83 @@ class SentryFlutterPlugin extends jni$_.JObject {
     return _privateSentryGetReplayIntegration(_class.reference.pointer,
             _id_privateSentryGetReplayIntegration as jni$_.JMethodIDPtr)
         .object<ReplayIntegration?>(const $ReplayIntegration$NullableType());
+  }
+
+  static final _id_getApplicationContext = _class.staticMethodId(
+    r'getApplicationContext',
+    r'()Landroid/content/Context;',
+  );
+
+  static final _getApplicationContext = jni$_.ProtectedJniExtensions.lookup<
+          jni$_.NativeFunction<
+              jni$_.JniResult Function(
+                jni$_.Pointer<jni$_.Void>,
+                jni$_.JMethodIDPtr,
+              )>>('globalEnv_CallStaticObjectMethod')
+      .asFunction<
+          jni$_.JniResult Function(
+            jni$_.Pointer<jni$_.Void>,
+            jni$_.JMethodIDPtr,
+          )>();
+
+  /// from: `static public final android.content.Context getApplicationContext()`
+  /// The returned object must be released after use, by calling the [release] method.
+  static jni$_.JObject? getApplicationContext() {
+    return _getApplicationContext(_class.reference.pointer,
+            _id_getApplicationContext as jni$_.JMethodIDPtr)
+        .object<jni$_.JObject?>(const jni$_.JObjectNullableType());
+  }
+
+  static final _id_loadContextsAsBytes = _class.staticMethodId(
+    r'loadContextsAsBytes',
+    r'()[B',
+  );
+
+  static final _loadContextsAsBytes = jni$_.ProtectedJniExtensions.lookup<
+          jni$_.NativeFunction<
+              jni$_.JniResult Function(
+                jni$_.Pointer<jni$_.Void>,
+                jni$_.JMethodIDPtr,
+              )>>('globalEnv_CallStaticObjectMethod')
+      .asFunction<
+          jni$_.JniResult Function(
+            jni$_.Pointer<jni$_.Void>,
+            jni$_.JMethodIDPtr,
+          )>();
+
+  /// from: `static public final byte[] loadContextsAsBytes()`
+  /// The returned object must be released after use, by calling the [release] method.
+  static jni$_.JByteArray? loadContextsAsBytes() {
+    return _loadContextsAsBytes(_class.reference.pointer,
+            _id_loadContextsAsBytes as jni$_.JMethodIDPtr)
+        .object<jni$_.JByteArray?>(const jni$_.JByteArrayNullableType());
+  }
+
+  static final _id_loadDebugImagesAsBytes = _class.staticMethodId(
+    r'loadDebugImagesAsBytes',
+    r'(Ljava/util/Set;)[B',
+  );
+
+  static final _loadDebugImagesAsBytes = jni$_.ProtectedJniExtensions.lookup<
+              jni$_.NativeFunction<
+                  jni$_.JniResult Function(
+                      jni$_.Pointer<jni$_.Void>,
+                      jni$_.JMethodIDPtr,
+                      jni$_.VarArgs<(jni$_.Pointer<jni$_.Void>,)>)>>(
+          'globalEnv_CallStaticObjectMethod')
+      .asFunction<
+          jni$_.JniResult Function(jni$_.Pointer<jni$_.Void>,
+              jni$_.JMethodIDPtr, jni$_.Pointer<jni$_.Void>)>();
+
+  /// from: `static public final byte[] loadDebugImagesAsBytes(java.util.Set<java.lang.String> set)`
+  /// The returned object must be released after use, by calling the [release] method.
+  static jni$_.JByteArray? loadDebugImagesAsBytes(
+    jni$_.JSet<jni$_.JString?> set,
+  ) {
+    final _$set = set.reference;
+    return _loadDebugImagesAsBytes(_class.reference.pointer,
+            _id_loadDebugImagesAsBytes as jni$_.JMethodIDPtr, _$set.pointer)
+        .object<jni$_.JByteArray?>(const jni$_.JByteArrayNullableType());
   }
 }